fixed font handling
due to inproper encoding, the web browser ie chrome had errors decoding the fonts as it always
displayed the `missaligned table and error decoding font` error on some fonts.
added an additional function that only re-encodes the data fetched from the file system to utf8 if the file is not a font file.
so as to ensure the font files are preserved and sent exactly as they where written to avoid data corruption.